add CustomTypes

+++ greycat/src/main/java/greycat/Container.java
@@ -68,7 +68,7 @@ public interface Container {
 
     Object getRawAt(int index);
 
-    Object getTypedRawAt(int index, byte type);
+    Object getTypedRawAt(int index, int type);
 
     /**
      * Returns the type of an attribute. The returned value is one of {@link Type}.
@@ -76,9 +76,9 @@ public interface Container {
      * @param name The name of the attribute for which the type is asked.
      * @return The type of the attribute inform of an int belonging to {@link Type}.
      */
-    byte type(String name);
+    int type(String name);
 
-    byte typeAt(int index);
+    int typeAt(int index);
 
     /**
      * Sets the value of an attribute of this container (for its current world and time for Node container).<br>
@@ -88,7 +88,7 @@ public interface Container {
      * @param value Must be consistent with the propertyType.
      * @return The node for fluent API.
      */
-    Container set(String name, byte type, Object value);
+    Container set(String name, int type, Object value);
 
     /**
      * Sets the value of an attribute of this container (for its current world and time for Node container).<br>
@@ -98,7 +98,7 @@ public interface Container {
      * @param value Must be consistent with the propertyType.
      * @return The node for fluent API.
      */
-    Container setAt(int index, byte type, Object value);
+    Container setAt(int index, int type, Object value);
 
     /**
      * Removes an attribute from the container.
@@ -117,7 +117,7 @@ public interface Container {
      * @param type The type of the attribute. Must be one of {@link Type} int value.
      * @return An instance that can be altered at the current world and time.
      */
-    Object getOrCreate(String name, byte type);
+    Object getOrCreate(String name, int type);
 
     /**
      * Gets or creates atomically a complex mutable attribute (e.g. Maps).<br>
@@ -126,8 +126,8 @@ public interface Container {
      * @param type  The type of the attribute. Must be one of {@link Type} int value.
      * @return An instance that can be altered at the current world and time.
      */
-    Object getOrCreateAt(int index, byte type);
-
+    Object getOrCreateAt(int index, int type);
+    
     <A> A getWithDefault(String key, A defaultValue);
 
     <A> A getAtWithDefault(int key, A defaultValue);
