+++ greycat/src/main/java/greycat/internal/tree/KDTreeNode.java
@@ -45,7 +45,7 @@ public class KDTreeNode extends BaseNode implements Tree {
     }
 
     @Override
-    public final Node set(String name, byte type, Object value) {
+    public final Node set(String name, int type, Object value) {
         if (name.equals(BOUND_MIN)) {
             setMinBound((double[]) value);
         } else if (name.equals(BOUND_MAX)) {
