+++ greycat/src/main/java/greycat/struct/proxy/ENodeProxy.java
@@ -134,12 +134,12 @@ public final class ENodeProxy implements ENode {
     }
 
     @Override
-    public final Object getOrCreate(final String name, final byte type) {
+    public final Object getOrCreate(final String name, final int type) {
         return this.getOrCreateAt(_node.egraph().graph().resolver().stringToHash(name, false), type);
     }
 
     @Override
-    public final Object getOrCreateAt(final int index, final byte type) {
+    public final Object getOrCreateAt(final int index, final int type) {
         Object elem = getAt(index);
         if (elem != null) {
             return proxifyIfNeeded(elem, index);
@@ -153,7 +153,7 @@ public final class ENodeProxy implements ENode {
         if (elem == null || _parent == null) { //implement time sensitivity
             return elem;
         } else {
-            byte type = typeAt(index);
+            int type = typeAt(index);
             switch (type) {
                 case Type.LMATRIX:
                     return new LMatrixProxy(index, this, (LMatrix) elem);
@@ -200,17 +200,17 @@ public final class ENodeProxy implements ENode {
     }
 
     @Override
-    public final Object getTypedRawAt(final int index, final byte type) {
+    public final Object getTypedRawAt(final int index, final int type) {
         return _node.getTypedRawAt(index, type);
     }
 
     @Override
-    public final byte type(final String name) {
+    public final int type(final String name) {
         return _node.type(name);
     }
 
     @Override
-    public final byte typeAt(final int index) {
+    public final int typeAt(final int index) {
         return _node.typeAt(index);
     }
 
@@ -229,13 +229,13 @@ public final class ENodeProxy implements ENode {
     }
 
     @Override
-    public final Container set(final String name, final byte type, final Object value) {
+    public final Container set(final String name, final int type, final Object value) {
         check();
         return _node.set(name, type, value);
     }
 
     @Override
-    public final Container setAt(final int index, final byte type, final Object value) {
+    public final Container setAt(final int index, final int type, final Object value) {
         check();
         return _node.setAt(index, type, value);
     }
